+++ declaration.d.ts
@@ -7,10 +7,6 @@ declare module "webp-converter" {
     ): Promise<void>;
 }
 
-declare module 'object-path' { 
-    export function get(item: any, path: string): any;
-}
-
 declare module 'md5' {
     export default function md5(input: string): string;
 }